requirements.txt
click>=8.1.3
google-generativeai>=0.7.2
